+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `CIVIC_REC_COMPLETE_CHECKOUT` schema to include `environmentId`
+
 ## [12.2.0] - 2025-09-19
 
 ### Added